# src/main/java/com/redhat/ecosystemappeng/rest/VulnerabilityEndpoint.java
package com.redhat.ecosystemappeng.rest;

import java.util.List;

import com.redhat.ecosystemappeng.model.Vulnerability;
import com.redhat.ecosystemappeng.service.VulnerabilityService;

import jakarta.inject.Inject;
import jakarta.ws.rs.GET;
import jakarta.ws.rs.POST;
import jakarta.ws.rs.Path;
import jakarta.ws.rs.PathParam;
import jakarta.ws.rs.QueryParam;

@Path("/vulnerabilities")
public class VulnerabilityEndpoint {

    @Inject
    VulnerabilityService cveService;

    @POST
    public List<Vulnerability> find(List<String> vulnerabilities, @QueryParam("reload") boolean reload) {
        return cveService.findByAlias(vulnerabilities, reload);
    }

    @GET
    @Path("/{alias}")
    public Vulnerability get(@PathParam("alias") String alias, @QueryParam("reload") boolean reload) {
        return cveService.getByAlias(alias, reload);
    }


}